split pinentry
